This drops some changes to files, which we will want to recreate namely:
 - plugins/gradle/tooling-extension-impl/src/com/intellij/gradle/toolingExtension/impl/model/taskModel/GradleTaskModelBuilder.kt
 - plugins/gradle/tooling-extension-impl/src/com/intellij/gradle/toolingExtension/impl/model/warmUp/GradleTaskWarmUpService.java
 - platform/build-scripts/src/org/jetbrains/intellij/build/impl/sbom/SoftwareBillOfMaterialsImpl.kt
 - platform/ide-core-impl/src/com/intellij/ide/impl/DataValidators.java
 - BUILD.bazel

+++ platform/platform-impl/src/com/intellij/execution/process/window/to/foreground/X11BringProcessWindowToForegroundSupport.kt
@@ -17,8 +17,8 @@ internal class X11BringProcessWindowToForegroundSupport : BringProcessWindowToFo
     }
   }
 
-  override fun bring(pid: Int): Boolean {
-    val mainWindow = X11UiUtil.findVisibleWindowByPid(pid) ?: run {
+  override fun bring(pid: UInt): Boolean {
+    val mainWindow = X11UiUtil.findVisibleWindowByPid(pid.toLong()) ?: run {
       logger.trace { "No window in \"$pid\" process found" }
       return false
     }
